requirements.txt
accelerate==1.6.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
antlr4-tools==0.2.1
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bitsandbytes==0.42.0
certifi==2025.1.31
charset-normalizer==3.4.1
chromadb-client==1.0.6
click==8.1.8
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
faiss-cpu==1.10.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.13.1
frozenlist==1.6.0
fsspec==2024.6.1
googleapis-common-protos==1.70.0
gradio==5.25.2
gradio_client==1.8.0
greenlet==3.2.0
groovy==0.1.2
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
install-jdk==1.1.0
interchange==2021.0.4
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.55
langchain-text-splitters==0.3.8
langsmith==0.3.33
mal-toolbox==0.3.11
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
opentelemetry-api==1.32.1
opentelemetry-exporter-otlp-proto-common==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-proto==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pansi==2024.11.0
peft==0.15.2
pillow==11.0.0
posthog==3.25.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
py2neo==2021.2.4
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.24.0
ruff==0.11.6
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
semantic-version==2.10.0
sentence-transformers==4.1.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
stix2==3.0.1
stix2-patterns==2.0.0
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zstandard==0.23.0
